Adjusted deltaTime handling
- DT uses double now
- Moved delta time update to the SceneDisplay
- Added a sleep_for to clamp to a 60 FPS frame rate

+++ Feather-Core/src/Core/CoreUtils/CoreEngineData.h
@@ -17,7 +17,7 @@ namespace Feather {
 		const float MetersToPixels() const;
 		const float PixelsToMeters() const;
 
-		inline float GetDeltaTime() const { return m_DeltaTime; }
+		inline double GetDeltaTime() const { return m_DeltaTime; }
 		inline int WindowWidth() const { return m_WindowWidth; }
 		inline int WindowHeight() const { return m_WindowHeight; }
 
@@ -56,7 +56,7 @@ namespace Feather {
 		CoreEngineData& operator=(const CoreEngineData&) = delete;
 
 	private:
-		float m_DeltaTime;
+		double m_DeltaTime;
 		float m_ScaledWidth;
 		float m_ScaledHeight;
 		float m_Gravity;
